LayoutTests/media/video-width-height.html

<video width=100 height=50 controls></video>
<script src=video-test.js></script>
<script>
test("video.width == 100");
test("video.height == 50");
test("video.offsetWidth == 100");
test("video.offsetHeight == 50");
video.setAttribute('width', '200');
video.setAttribute('height', '100');
test("video.width == 200");
test("video.height == 100");
test("video.offsetWidth == 200");
test("video.offsetHeight == 100");
video.width = 400;
video.height = 200;
test("video.getAttribute('width') == '400'");
test("video.getAttribute('height') == '200'");
test("video.offsetWidth == 400");
test("video.offsetHeight == 200");
endTest();
</script>
