LayoutTests/media/video-width-height-expected.txt

TEST(video.width == 100) OK
TEST(video.height == 50) OK
TEST(video.offsetWidth == 100) OK
TEST(video.offsetHeight == 50) OK
TEST(video.width == 200) OK
TEST(video.height == 100) OK
TEST(video.offsetWidth == 200) OK
TEST(video.offsetHeight == 100) OK
TEST(video.getAttribute('width') == '400') OK
TEST(video.getAttribute('height') == '200') OK
TEST(video.offsetWidth == 400) OK
TEST(video.offsetHeight == 200) OK
END OF TEST
